feat(docker): Dockerize the flask application
Create docker image, run the container and test the APIs on Postman.
Modify flask app db access: Use 'host.docker.internal' instead of
'localhost' as localhost would mean localhost of the container and
not that of the host. Use 0.0.0.0 as host to make the container
publicly available in the same network.